requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pypdf==4.3.1
defusedxml==0.7.1
pytest==8.3.2
